// apps/finance-web/src/models/product/ParameterType.ts
export enum ParameterType {
  DICT = 'DICT',
  MULTI_DICT = 'MULTI_DICT',
  CALC = 'CALC',
  COMBO_DICT = 'COMBO_DICT',
  PERIOD = 'PERIOD',
  RANGE = 'RANGE',
  LONG = 'LONG',
  COMMISSION = 'COMMISSION',
  INT = 'INT',
  DOUBLE = 'DOUBLE',
  STR = 'STR',
  PERCENTAGE = 'PERCENTAGE',
  SELECTOR = 'SELECTOR',
  DATE_SELECTOR = 'DATE_SELECTOR',
  DESCRIPTION = 'DESCRIPTION',
  REGION_SELECTOR = 'REGION_SELECTOR',
  NOT_MULTI_DICT = 'NOT_MULTI_DICT',
  DICT_CREDIT_HISTORY = 'DICT_CREDIT_HISTORY',
}